WIP: Refactor server and state logic
