+++ src/state/state.rs
@@ -1,17 +1,21 @@
-#[derive(Clone, Copy)]
-struct Vec2 {
-    x: f32,
-    y: f32,
-}
+use super::model::*;
+use crate::game_state_generated::Color;
 
-pub struct PlayerState {
-    pos: Vec2,
-    vel: Vec2,
-    acc: f32,
-    jump_force: f32,
-    jump_timer: f32,
-    color: Color,
-    size: f32,
+use super::{JUMP_CD, SCREEN_HEIGHT};
+
+impl PlayerState {
+    fn new(id: u32) -> PlayerState {
+        PlayerState {
+            id,
+            pos: Vec2::zero(),
+            vel: Vec2::zero(),
+            acc: 0.75,
+            jump_force: 400.0,
+            jump_timer: 0.0,
+            color: Color::Red,
+            size: 16.0,
+        }
+    }
 }
 
 pub struct GameState {
@@ -19,5 +23,24 @@ pub struct GameState {
 }
 
 impl GameState {
-    pub fn mutate(&mut self, commands[])
+    pub fn mutate(&mut self) {
+        println!("Mutating!")
+    }
+}
+
+impl PlayerState {
+    fn handle_move_right(&mut self) {
+        self.vel.x += self.acc;
+    }
+
+    fn handle_move_left(&mut self) {
+        self.vel.x -= self.acc;
+    }
+
+    fn handle_jump(&mut self) {
+        if self.pos.y >= SCREEN_HEIGHT as f32 - self.size && self.jump_timer > JUMP_CD {
+            self.vel.y -= self.jump_force;
+            self.jump_timer = 0.0;
+        };
+    }
 }
